fixed specialAbs in desc with links; fixed edge with parenthesis
